# server/openapi.yaml
openapi: 3.0.0
info:
  title: Mini Supermarket API
  version: 1.0.0
  description: API cho hệ thống quản lý siêu thị mini
servers:
  - url: 'http://localhost:5000'
...
paths:
  /api/auth/register/customer:
    post:
      tags:
        - auth
      summary: register Customer
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/auth/register/staff:
    post:
      tags:
        - auth
      summary: register Staff
      operationId: registerStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/auth/login:
    post:
      tags:
        - auth
      summary: login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/auth/me:
    get:
      tags:
        - auth
      summary: get Me
      operationId: getMe
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/auth/update-profile:
    put:
      tags:
        - auth
      summary: update Profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/auth/change-password:
    put:
      tags:
        - auth
      summary: change Password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/auth/verify-token:
    post:
      tags:
        - auth
      summary: verify Token
      operationId: verifyToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...